♻️ rename message-text.tsx to text-message.tsx
