+++ app/components/message/text-message.tsx
